fastapi/requirements.txt
uvicorn 
fastapi
openai
tiktoken
python-dotenv